+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All Notable changes to `bakame/shoes-size-converter` will be documented in this file.
 
-## [0.1.0](https://github.com/bakame-php/shoes-size-converter/releases/tag/0.1.0) - 2025-08-20
+## [Next](https://github.com/bakame-php/shoes-size-converter/releases/tag/0.1.0) - 2026-08-20
+
+- Add general support for `TabularData` interface
+- Add `ShoeException` base exception
+- Rewrote the script to use any unit and size.
+
+## [0.1.0](https://github.com/bakame-php/shoes-size-converter/releases/tag/0.1.0) - 2026-08-20
 
 **Initial release!**